Move config to API module
